requirements.txt
selenium
langchain
openai
docarray
streamlit
unstructured
faiss-cpu
tiktoken
python-dotenv
beautifulsoup4